fix: add micrometer bom to observability & optimize the structure of the package (#11724)
* chore: optimize the structure of the package

* chore: add micrometer bom to observability

---------
