+++ dubbo-spring-boot/dubbo-spring-boot-observability-starter/pom.xml
@@ -28,6 +28,30 @@
 
     <artifactId>dubbo-spring-boot-observability-starter</artifactId>
 
+    <properties>
+        <micrometer.version>1.10.4</micrometer.version>
+        <micrometer-tracing.version>1.0.2</micrometer-tracing.version>
+    </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>io.micrometer</groupId>
+                <artifactId>micrometer-bom</artifactId>
+                <version>${micrometer.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+            <dependency>
+                <groupId>io.micrometer</groupId>
+                <artifactId>micrometer-tracing-bom</artifactId>
+                <version>${micrometer-tracing.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
     <dependencies>
         <dependency>
             <groupId>io.micrometer</groupId>
